requirements.txt
Django==5.0.6
mysqlclient==2.2.4
mysql-connector-python==8.4.0
djangorestframework==3.15.1
django-filter==24.2
djoser==2.2.2